More changes to especially the compilation, with the addition of the intel compiler

+++ FortranSerial/source/PrecisionMod.f90
@@ -5,8 +5,8 @@ module PrecMod
   use iso_fortran_env, only:  int8, int16, int32, int64, real32, real64
   implicit none
 
-  integer, parameter :: wpf = real32
-  integer, parameter :: wpi = int32
+  integer, parameter :: wpf = real64
+  integer, parameter :: wpi = int64
 
   
 end module PrecMod
